fix / Resolve issue of data loss upon page refresh
Router.jsx파일에서 파이어스토어를 fetch해서 새로고침시 데이터가 없어지는 문제를 해결함
