+++ src/components/home/HomeComp.jsx
@@ -10,43 +10,12 @@ const HomeComp = () => {
   const dispatch = useDispatch();
 
   useEffect(() => {
-    const newArr = [];
-    const fetchPostsData = async () => {
-      const q = query(collection(db, "posts"));
-      const querySnapshot = await getDocs(q);
-      querySnapshot.forEach((doc) => {
-        newArr.push({ id: doc.id, ...doc.data() });
-      });
-      dispatch(showPosts(newArr));
-      dispatch(sortLikePosts());
-    };
-    fetchPostsData();
-
-    const fetchMemberData = async () => {
-      // q = 요청 객체
-      const q = query(collection(db, "starList"));
-      const querySnapshot = await getDocs(q);
-      const initialStarList = [];
-      querySnapshot.forEach((doc) => {
-        const data = {
-          id: doc.id,
-          ...doc.data(),
-        };
-        initialStarList.push(data);
-      });
-      dispatch(showMembers(initialStarList));
-      dispatch(sortLikeMembers());
-    };
-    fetchMemberData();
+    dispatch(sortLikePosts());
+    dispatch(sortLikeMembers());
   }, []);
 
   const popPosts = useSelector((state) => state.posts);
-  console.log("popPosts", popPosts);
-
-  const StarList = useSelector((state) => {
-    return state.logReducer.members;
-  });
-  console.log("StarList", StarList.slice(0, 5));
+  const StarList = useSelector((state) => state.logReducer.members);
 
   return (
     <>
