+++ src/redux/modules/postWrite.js
@@ -37,7 +37,6 @@ const posts = (state = newArr, action) => {
         }
       });
     case SORT_LIKE_POSTS:
-      console.log("state", state);
       return state.sort((a, b) => b.postWhoLiked?.length || 0 - a.postWhoLiked?.length || 0);
     case "UPDATE_POSTLIKE":
       return state.map((post) => {
