+++ src/shared/Router.jsx
@@ -18,8 +18,12 @@ import PostUpdate from "../pages/PostUpdate";
 //진솔 추가
 import { onAuthStateChanged } from "firebase/auth";
 import { useDispatch } from "react-redux";
-import { showUser } from "../redux/modules/logReducer";
-import { auth } from "../firebase";
+import { showMembers, showUser, sortLikeMembers } from "../redux/modules/logReducer";
+import { auth, db } from "../firebase";
+
+import { useEffect } from "react";
+import { collection, getDocs, query } from "@firebase/firestore";
+import { showPosts, sortLikePosts } from "../redux/modules/postWrite";
 
 const Router = () => {
   const dispatch = useDispatch();
@@ -30,6 +34,34 @@ const Router = () => {
       // dispatch(logChange(true));
     } else return;
   });
+  useEffect(() => {
+    const newArr = [];
+    const fetchPostsData = async () => {
+      const q = query(collection(db, "posts"));
+      const querySnapshot = await getDocs(q);
+      querySnapshot.forEach((doc) => {
+        newArr.push({ id: doc.id, ...doc.data() });
+      });
+      dispatch(showPosts(newArr));
+    };
+    fetchPostsData();
+
+    const fetchMemberData = async () => {
+      // q = 요청 객체
+      const q = query(collection(db, "starList"));
+      const querySnapshot = await getDocs(q);
+      const initialStarList = [];
+      querySnapshot.forEach((doc) => {
+        const data = {
+          id: doc.id,
+          ...doc.data(),
+        };
+        initialStarList.push(data);
+      });
+      dispatch(showMembers(initialStarList));
+    };
+    fetchMemberData();
+  }, []);
 
   return (
     <>
